requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
numpy==1.17.0
pandas==0.25.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
urllib3==1.25.3
